+++ src/publishers.rs
@@ -4,42 +4,42 @@ use std::io::Result;
 
 #[derive(Deserialize)]
 struct UsersResponse {
-    users: Vec<OwnerData>,
+    users: Vec<PublisherData>,
 }
 
 #[derive(Deserialize)]
 struct TeamsResponse {
-    teams: Vec<OwnerData>,
+    teams: Vec<PublisherData>,
 }
 
 #[derive(Deserialize, Debug, Clone)]
-pub struct OwnerData {
+pub struct PublisherData {
     pub id: u64,
     pub login: String,
-    pub kind: OwnerKind,
+    pub kind: PublisherKind,
     pub url: Option<String>,
     pub name: Option<String>,
     pub avatar: Option<String>,
 }
 
-impl PartialEq for OwnerData {
+impl PartialEq for PublisherData {
     fn eq(&self, other: &Self) -> bool {
         self.id == other.id
     }
 }
 
-impl Eq for OwnerData {
-    // holds for OwnerData because we're comparing u64 IDs, and it holds for u64
+impl Eq for PublisherData {
+    // holds for PublisherData because we're comparing u64 IDs, and it holds for u64
     fn assert_receiver_is_total_eq(&self) {}
 }
 
-impl PartialOrd for OwnerData {
+impl PartialOrd for PublisherData {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
         Some(self.id.cmp(&other.id))
     }
 }
 
-impl Ord for OwnerData {
+impl Ord for PublisherData {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
         self.id.cmp(&other.id)
     }
@@ -47,26 +47,26 @@ impl Ord for OwnerData {
 
 #[derive(Deserialize, Debug, Copy, Clone)]
 #[allow(non_camel_case_types)]
-pub enum OwnerKind {
+pub enum PublisherKind {
     user,
     team,
 }
 
-pub fn owner_users(client: &mut ApiClient, crate_name: &str) -> Result<Vec<OwnerData>> {
-    let url = format!("https://crates.io/api/v1/crates/{}/owner_user", crate_name);
+pub fn publisher_users(client: &mut ApiClient, crate_name: &str) -> Result<Vec<PublisherData>> {
+    let url = format!("https://crates.io/api/v1/crates/{}/publisher_user", crate_name);
     let data: UsersResponse = client.get(&url).call().into_json_deserialize()?;
     Ok(data.users)
 }
 
-pub fn owner_teams(client: &mut ApiClient, crate_name: &str) -> Result<Vec<OwnerData>> {
-    let url = format!("https://crates.io/api/v1/crates/{}/owner_team", crate_name);
+pub fn publisher_teams(client: &mut ApiClient, crate_name: &str) -> Result<Vec<PublisherData>> {
+    let url = format!("https://crates.io/api/v1/crates/{}/publisher_team", crate_name);
     let data: TeamsResponse = client.get(&url).call().into_json_deserialize()?;
     Ok(data.teams)
 }
 
-pub fn owners(client: &mut ApiClient, crate_name: &str) -> Result<Vec<OwnerData>> {
-    let mut users = owner_users(client, crate_name)?;
-    let mut teams = owner_teams(client, crate_name)?;
+pub fn publishers(client: &mut ApiClient, crate_name: &str) -> Result<Vec<PublisherData>> {
+    let mut users = publisher_users(client, crate_name)?;
+    let mut teams = publisher_teams(client, crate_name)?;
     users.extend(teams.drain(..));
     Ok(users)
 }