+++ Changes.md
@@ -5,4 +5,4 @@ Initial release, supports one command:
   Authors might be listed multiple times. For each author, differentiate if
   they are known by being mentioned in a crate from the local workspace or not.
   Support for crawling `crates.io` sourced packages is planned.
-- `owners`: Doesn't do anything right now.
+- `publishers`: Doesn't do anything right now.